CAUSEWAY-3597: moves downloadColumnOrderTxtFileAsZip mixin
to applib, introducing new domain service to enable this
