+++ extensions/vw/pdfjs/metamodel/src/test/java/org/apache/causeway/extensions/pdfjs/metamodel/PdfjsViewer_PropDomain_IntegTest.dump_facets.approved.xml
@@ -200,7 +200,7 @@
             <mml:attr name="precedence" value="DEFAULT"/>
           </mml:facet>
           <mml:facet id="org.apache.causeway.core.metamodel.facets.actions.action.invocation.ActionDomainEventFacet" fqcn="org.apache.causeway.core.metamodel.facets.actions.action.invocation.ActionDomainEventFacet">
-            <mml:attr name="eventType" value="org.apache.causeway.core.metamodel.inspect.Object_downloadColumnOrderTxtFilesAsZip.ActionDomainEvent"/>
+            <mml:attr name="eventType" value="org.apache.causeway.applib.services.columnorder.Object_downloadColumnOrderTxtFilesAsZip.ActionDomainEvent"/>
             <mml:attr name="eventTypeOrigin" value="ANNOTATED_MEMBER"/>
             <mml:attr name="facet" value="ActionDomainEventFacet"/>
             <mml:attr name="interactionAdvisors" value="Validating, HidingOrShowing, DisablingOrEnabling"/>
@@ -208,13 +208,13 @@
             <mml:attr name="precedence" value="DEFAULT"/>
           </mml:facet>
           <mml:facet id="org.apache.causeway.core.metamodel.facets.actions.action.invocation.ActionInvocationFacet" fqcn="org.apache.causeway.core.metamodel.facets.actions.action.invocation.ActionInvocationFacetForAction">
-            <mml:attr name="declaringType" value="org.apache.causeway.core.metamodel.inspect.Object_downloadColumnOrderTxtFilesAsZip"/>
-            <mml:attr name="eventType" value="org.apache.causeway.core.metamodel.inspect.Object_downloadColumnOrderTxtFilesAsZip.ActionDomainEvent"/>
+            <mml:attr name="declaringType" value="org.apache.causeway.applib.services.columnorder.Object_downloadColumnOrderTxtFilesAsZip"/>
+            <mml:attr name="eventType" value="org.apache.causeway.applib.services.columnorder.Object_downloadColumnOrderTxtFilesAsZip.ActionDomainEvent"/>
             <mml:attr name="eventTypeOrigin" value="ANNOTATED_MEMBER"/>
             <mml:attr name="facet" value="ActionInvocationFacetForAction"/>
             <mml:attr name="intent.act" value="EXECUTE"/>
             <mml:attr name="isPostable" value="true"/>
-            <mml:attr name="methods" value="public org.apache.causeway.applib.value.Blob org.apache.causeway.core.metamodel.inspect.Object_downloadColumnOrderTxtFilesAsZip.act(java.lang.String)"/>
+            <mml:attr name="methods" value="public org.apache.causeway.applib.value.Blob org.apache.causeway.applib.services.columnorder.Object_downloadColumnOrderTxtFilesAsZip.act(java.lang.String)"/>
             <mml:attr name="precedence" value="DEFAULT"/>
             <mml:attr name="returnType" value="org.apache.causeway.applib.value.Blob"/>
           </mml:facet>
@@ -340,7 +340,7 @@
               <mml:facet id="org.apache.causeway.core.metamodel.facets.param.defaults.ActionParameterDefaultsFacet" fqcn="org.apache.causeway.core.metamodel.facets.param.defaults.methodnum.ActionParameterDefaultsFacetViaMethod">
                 <mml:attr name="facet" value="ActionParameterDefaultsFacetViaMethod"/>
                 <mml:attr name="intent.default0Act" value="DEFAULTS"/>
-                <mml:attr name="methods" value="public java.lang.String org.apache.causeway.core.metamodel.inspect.Object_downloadColumnOrderTxtFilesAsZip.default0Act()"/>
+                <mml:attr name="methods" value="public java.lang.String org.apache.causeway.applib.services.columnorder.Object_downloadColumnOrderTxtFilesAsZip.default0Act()"/>
                 <mml:attr name="precedence" value="DEFAULT"/>
               </mml:facet>
               <mml:facet id="org.apache.causeway.core.metamodel.facets.param.parameter.depdef.ParameterDependentDefaultsFacet" fqcn="org.apache.causeway.core.metamodel.facets.param.parameter.depdef.ParameterDependentDefaultsFacetFromConfiguration">
